started portfolios page functionality
