+++ src/main/web-static/websrc/infra/routing/binding-dom-navigate-to.ts
@@ -1,4 +1,4 @@
-import { navigoRouter } from "./routing-navigo";
+import { buildParameterizedDestinationPath, navigoRouter } from "./routing-navigo";
 import { logger, LogLevel } from "../logging";
 import DomUtils from "../dom/dom-utils";
 
@@ -9,11 +9,6 @@ import DomUtils from "../dom/dom-utils";
 
 const NAVIGATE_TO_ATTRIBUTE = "data-navigate-to";
 const NAVIGATE_TO_BOUND_FLAG = "navigate-to-bound";
-const NAVIGO_PATH_PARAM_PREFIX = ":";
-
-function hasPathParam(path: string) {
-    return path.includes(NAVIGO_PATH_PARAM_PREFIX);
-}
 
 function bindKeypressNavigation(element: HTMLElement) {
     element.addEventListener("keypress", (event: KeyboardEvent) => {
@@ -35,35 +30,8 @@ function navigate(element: HTMLElement) {
 }
 
 function buildDestinationPath(element: HTMLElement) {
-
-    let destinationPath = element.getAttribute(NAVIGATE_TO_ATTRIBUTE);
-
-    if(hasPathParam(destinationPath)) {
-        destinationPath = resolvePathParamsFromCurrentLocationContext(destinationPath);
-    }
-
-    return destinationPath;
-}
-
-function resolvePathParamsFromCurrentLocationContext(path: string) {
-
-    let resolvedPath = path;
-    const currentLocation = navigoRouter.getCurrentLocation();
-    const currentMatch = navigoRouter.match(currentLocation.url)[0];
-
-    if(currentMatch) {
-        for(const key in currentMatch.data) {
-            resolvedPath = resolvedPath.replace(`${ NAVIGO_PATH_PARAM_PREFIX }${ key }`, currentMatch.data[key]);
-        }
-    }
-
-    if(hasPathParam(resolvedPath)) {
-        const errorMessage = "Could not find path param in current route navigate to";
-        logger(LogLevel.ERROR, errorMessage, resolvedPath);
-        throw new Error(errorMessage);
-    }
-
-    return resolvedPath;
+    const destinationPath = element.getAttribute(NAVIGATE_TO_ATTRIBUTE);
+    return buildParameterizedDestinationPath(destinationPath);
 }
 
 function bindNavigateToElements(navigationElement: NodeListOf<HTMLElement>) {
