+++ src/main/web-static/websrc/infra/routing/router.ts
@@ -2,6 +2,8 @@ import { bindHTMXTriggerOnRouteInDescendants } from "./binding-htmx-trigger-on-r
 import { bindNavigateToInDescendants } from "./binding-dom-navigate-to";
 import { bindDisplayOnRouteInDescendants } from "./binding-dom-display-on-route";
 import { bindAttributeOnRouteInDescendants } from "./binding-dom-attribute-on-route";
+import { bootNavigoRouter } from "./routing-navigo";
+
 
 const router = {
     init() {
@@ -15,6 +17,9 @@ const router = {
     },
     bindDocumentToRouting() {
         this.bindDescendants(document.body);
+    },
+    boot() {
+        bootNavigoRouter();
     },
 };
